items

+++ scripts/macros/items/trinket/allPurposeTool.js
@@ -1,42 +1,44 @@
-import {dialogUtils, effectUtils, genericUtils, itemUtils, spellUtils, workflowUtils} from '../../../utils.js';
+import {activityUtils, dialogUtils, effectUtils, genericUtils, itemUtils, spellUtils, workflowUtils} from '../../../utils.js';
 async function use({trigger, workflow}) {
     if (!itemUtils.getEquipmentState(workflow.item)) return;
-    let key = genericUtils.getCPRSetting('itemCompendium');
-    if (!key || key === '') return;
-    let pack = game.packs.get(key);
-    if (!pack) return;
-    let index = await pack.getIndex({fields: ['system.type.value']});
-    let tools = index.filter(i => i.system.type?.value === 'art');
-    let tool = await dialogUtils.selectDocumentDialog(workflow.item.name, 'CHRISPREMADES.Macros.AllPurposeTool.Select', tools, {sortAlphabetical: true});
-    if (!tool) return;
-    let oldTool = workflow.actor.items.find(i => i.flags['chris-premades']?.allPurposeTool);
-    if (oldTool) await oldTool.delete();
-    let itemData = genericUtils.duplicate(tool.toObject());
-    genericUtils.setProperty(itemData, 'flags.chris-premades.allPurposeTool', true);
-    itemData.name = workflow.item.name + ': ' + itemData.name;
-    let prof = itemUtils.getToolProficiency(workflow.actor, tool);
-    if (!prof) genericUtils.setProperty(itemData, 'system.proficient', 1);
-    await itemUtils.createItems(workflow.actor, [itemData], {parentEntity: workflow.item});
-}
-async function spell({trigger, workflow}) {
-    let spells = await spellUtils.getSpellsOfLevel(0);
-    let selection = await dialogUtils.selectDocumentDialog(workflow.item.name, 'CHRISPREMADES.Macros.AllPurposeTool.Spell', spells, {sortAlphabetical: true});
-    if (!selection) return;
-    let itemData = duplicate(selection.toObject());
-    genericUtils.setProperty(itemData, 'system.sourceClass', 'artificer');
-    itemData.system.properties.push('material');
-    itemData.name = genericUtils.translate('CHRISPREMADES.Macros.AllPurposeTool.Name') + ': ' + itemData.name;
-    let effectData = {
-        name: workflow.item.name,
-        img: workflow.item.img,
-        origin: workflow.item.uuid,
-        duration: itemUtils.convertDuration(workflow.activity)
-    };
-    let effect = await effectUtils.createEffect(workflow.actor, effectData, {parentEntity: workflow.item});
-    await itemUtils.createItems(workflow.actor, [itemData], {parentEntity: effect});
+    let activityIdentifier = activityUtils.getIdentifier(workflow.activity);
+    if (activityIdentifier === 'allPurposeTool') {
+        let key = genericUtils.getCPRSetting('itemCompendium');
+        if (!key || key === '') return;
+        let pack = game.packs.get(key);
+        if (!pack) return;
+        let index = await pack.getIndex({fields: ['system.type.value']});
+        let tools = index.filter(i => i.system.type?.value === 'art');
+        let tool = await dialogUtils.selectDocumentDialog(workflow.item.name, 'CHRISPREMADES.Macros.AllPurposeTool.Select', tools, {sortAlphabetical: true});
+        if (!tool) return;
+        let oldTool = workflow.actor.items.find(i => i.flags['chris-premades']?.allPurposeTool);
+        if (oldTool) await oldTool.delete();
+        let itemData = genericUtils.duplicate(tool.toObject());
+        genericUtils.setProperty(itemData, 'flags.chris-premades.allPurposeTool', true);
+        itemData.name = workflow.item.name + ': ' + itemData.name;
+        let prof = itemUtils.getToolProficiency(workflow.actor, tool);
+        if (!prof) genericUtils.setProperty(itemData, 'system.proficient', 1);
+        await itemUtils.createItems(workflow.actor, [itemData], {parentEntity: workflow.item});
+    } else if (activityIdentifier === 'channelCreativeForces') {
+        let spells = await spellUtils.getSpellsOfLevel(0);
+        let selection = await dialogUtils.selectDocumentDialog(workflow.item.name, 'CHRISPREMADES.Macros.AllPurposeTool.Spell', spells, {sortAlphabetical: true});
+        if (!selection) return;
+        let itemData = duplicate(selection.toObject());
+        genericUtils.setProperty(itemData, 'system.sourceClass', 'artificer');
+        itemData.system.properties.push('material');
+        itemData.name = genericUtils.translate('CHRISPREMADES.Macros.AllPurposeTool.Name') + ': ' + itemData.name;
+        let effectData = {
+            name: workflow.item.name,
+            img: workflow.item.img,
+            origin: workflow.item.uuid,
+            duration: itemUtils.convertDuration(workflow.activity)
+        };
+        let effect = await effectUtils.createEffect(workflow.actor, effectData, {parentEntity: workflow.item});
+        await itemUtils.createItems(workflow.actor, [itemData], {parentEntity: effect});
+    }
 }
 async function early({trigger: {entity: item}, workflow}) {
-    if (!workflow.item.hasSave) return;
+    if (!workflow.activity.save) return;
     if (!spellUtils.isClassSpell(workflow.item, 'artificer')) return;
     let identifier = genericUtils.getIdentifier(item);
     let bonus;
@@ -48,10 +50,13 @@ async function early({trigger: {entity: item}, workflow}) {
         bonus = 3;
     }
     if (!bonus) return;
-    workflow.item = workflow.item.clone({'system.save.dc': workflow.item.system.save.dc + bonus, 'system.save.scaling': 'flat'}, {keepId: true});
-    workflow.item.prepareData();
-    workflow.item.prepareFinalAttributes();
-    workflow.item.applyActiveEffects();
+    let newActivity = genericUtils.deepClone(workflow.activity);
+    newActivity.save.dc = {
+        calculation: '',
+        formula: workflow.activity.save.dc.formula + ' + ' + bonus,
+        value: workflow.activity.save.dc.value + bonus
+    };
+    workflow.activity = newActivity;
 }
 async function attack({trigger: {entity: item}, workflow}) {
     if (!spellUtils.isClassSpell(workflow.item, 'artificer')) return;
@@ -69,7 +74,7 @@ async function attack({trigger: {entity: item}, workflow}) {
 }
 export let allPurposeTool1 = {
     name: 'All Purpose Tool + 1',
-    version: '1.0.37',
+    version: '1.1.0',
     midi: {
         item: [
             {
@@ -90,49 +95,15 @@ export let allPurposeTool1 = {
                 priority: 50
             },
         ]
-    },
-    equipment: {
-        channelCreativeForces: {
-            name: 'All-Purpose Tool: Channel Creative Forces',
-            compendium: 'itemEquipment',
-            useJournal: true,
-            uses: {
-                max: 1,
-                recovery: [
-                    {
-                        period: 'dawn',
-                        type: 'recoverAll'
-                    }
-                ],
-                spent: 0
-            },
-            translate: 'CHRISPREMADES.Macros.AllPurposeTool.Channel',
-            favorite: true
-        }
     }
 };
 export let allPurposeTool2 = {
     name: 'All Purpose Tool + 2',
     version: allPurposeTool1.version,
-    midi: allPurposeTool1.midi,
-    equipment: allPurposeTool1.equipment
+    midi: allPurposeTool1.midi
 };
 export let allPurposeTool3 = {
     name: 'All Purpose Tool + 3',
     version: allPurposeTool1.version,
-    midi: allPurposeTool1.midi,
-    equipment: allPurposeTool1.equipment
-};
-export let channelCreativeForces = {
-    name: 'All-Purpose Tool: Channel Creative Forces',
-    version: allPurposeTool1.version,
-    midi: {
-        item: [
-            {
-                pass: 'rollFinished',
-                macro: spell,
-                priority: 50
-            }
-        ]
-    }
+    midi: allPurposeTool1.midi
 };
